refactor:#118 캘린더 페이지와 컴포넌트 역할 재설정- 페이지:약속 데이터 생성/컴포넌트:전달받은 데이터로 그려주기
