+++ src/components/calendar/MainCalendar.tsx
@@ -1,17 +1,14 @@
 import { Calendar, dateFnsLocalizer } from 'react-big-calendar';
 import { format, parse, startOfWeek, getDay } from 'date-fns';
-import { useGetCalendarPlans } from '@/hooks/queries/useGetCalendarPlans';
 import CustomToolbar from '@/components/calendar/CustomToolbar';
 import { useEffect, useState } from 'react';
-import type { CalendarEventType, Holidays, PlansType } from '@/types/plans';
+import type { CalendarEventType, PlansType } from '@/types/plans';
 import withDragAndDrop from 'react-big-calendar/lib/addons/dragAndDrop';
 import { useUpadateEventMutate } from '@/hooks/mutations/useUpadateEventMutate';
 import { CustomDateHeader } from '@/components/calendar/CustomDateHeader';
-import { useGetHolidays } from '@/hooks/queries/useGetHolidays';
 import { holidayStyle } from '@/lib/utils/calendarStyle';
 import CalendarPopOver from '@/components/calendar/popOver/CalendarPopOver';
 import { useGetSelectPlan } from '@/hooks/queries/useGetSelectPlan';
-import { User } from '@supabase/supabase-js';
 
 // 드래그 이벤트 타입
 interface Props {
@@ -21,7 +18,9 @@ interface Props {
 }
 
 interface MainCalendarProps {
-  user: User | null;
+  events: CalendarEventType[];
+  moment: Date;
+  setMoment: (date: Date) => void;
   setSelectPlan: React.Dispatch<React.SetStateAction<PlansType[] | null>>;
   CustomToolbarProps: {
     onShowUpcomingPlans: () => void;
@@ -37,31 +36,13 @@ const localizer = dateFnsLocalizer({
   locales: {},
 });
 
-const MainCalendar = ({ setSelectPlan, CustomToolbarProps, user }: MainCalendarProps) => {
-  const [moment, setMoment] = useState(new Date()); //해당 달
+const MainCalendar = ({ setSelectPlan, CustomToolbarProps, events, moment, setMoment }: MainCalendarProps) => {
   const [selectedDate, setSelectedDate] = useState<Date | null>(null); //선택한 셀 날짜
   const [isPopOverOpen, setIsPopOverOpen] = useState(false); //팝오버 오픈 여부
   const [selectedPlanId, setSelectedPlanId] = useState<string | null>(null);
 
-  const calendarYear = moment.getFullYear(); // 해당 달의 년도
-
   const { data: selectedPlanData, refetch } = useGetSelectPlan(selectedPlanId ?? '');
   const { mutate: updateEvent } = useUpadateEventMutate();
-  const { data: holidays } = useGetHolidays(String(calendarYear));
-  const { data: events, isPending, isError, error } = useGetCalendarPlans(user, calendarYear, moment);
-
-  // 약속 + 공휴일
-  const combinedEvents: CalendarEventType[] = [
-    ...(events || []),
-    ...(holidays || []).map((holiday: Holidays, idx: number) => ({
-      id: `holiday-${idx}`,
-      title: holiday.title,
-      start: holiday.date,
-      end: holiday.date, // 단일 일정
-      isHoliday: true, // 스타일 구분용
-      colors: '#2F80ED', // 기본 색상
-    })),
-  ];
 
   const DnDCalendar = withDragAndDrop<CalendarEventType>(Calendar); //DnD 사용 캘린더
 
@@ -95,20 +76,12 @@ const MainCalendar = ({ setSelectPlan, CustomToolbarProps, user }: MainCalendarP
     }
   }, [selectedPlanId, refetch]);
 
-  if (isPending) {
-    return <div>로딩 중입니다...</div>;
-  }
-
-  if (isError) {
-    return <div>캘린더 에러 발생 : {error.message}</div>;
-  }
-
   return (
     <div>
       <DnDCalendar
         selectable
         localizer={localizer}
-        events={combinedEvents}
+        events={events}
         date={moment} // 현재 달 state
         onNavigate={(newDate) => {
           setMoment(newDate); // 달 변동
